add EnterGameRoom
