+++ .metadata/.plugins/org.eclipse.ui.workbench.texteditor/dialog_settings.xml
@@ -1,13 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <section name="Workbench">
 	<section name="org.eclipse.ui.texteditor.FindReplaceDialog">
-		<item key="selection" value=""/>
+		<item key="selection" value="enterRoom"/>
 		<item key="wrap" value="true"/>
 		<item key="casesensitive" value="false"/>
 		<item key="wholeword" value="false"/>
 		<item key="incremental" value="false"/>
 		<item key="isRegEx" value="false"/>
 		<list key="findhistory">
+			<item value="enterRoom"/>
 			<item value="&quot;O&quot;"/>
 			<item value="WriteAllO"/>
 			<item value="Client get"/>
