+++ GameClient/src/client/main/GameClient.java
@@ -11,7 +11,7 @@ import javax.swing.JTextField;
 import javax.swing.SwingConstants;
 import javax.swing.border.EmptyBorder;
 
-import client.wait.WaitingView;
+import client.room.WaitingView;
 
 public class GameClient extends JFrame{
