+++ GameClient/src/data/ChatMsg.java
@@ -1,13 +1,16 @@
+// ChatMsg.java 채팅 메시지 ObjectStream 용.
 package data;
-import java.io.Serializable;
 
+import java.io.Serializable;
 import javax.swing.ImageIcon;
+import data.GameRoom;
 
 public class ChatMsg implements Serializable {
 	private static final long serialVersionUID = 1L;
 	private String id;
 	private String code;
 	private String data;
+	private GameRoom room;
 	public ImageIcon img;
 
 	public ChatMsg(String id, String code, String msg) {
@@ -31,6 +34,10 @@ public class ChatMsg implements Serializable {
 	public String getId() {
 		return id;
 	}
+	
+	public GameRoom getGameRoom() {
+		return room;
+	}
 
 	public void setId(String id) {
 		this.id = id;
@@ -43,5 +50,8 @@ public class ChatMsg implements Serializable {
 	public void setImg(ImageIcon img) {
 		this.img = img;
 	}
-
+	
+	public void setRoom(GameRoom room) {
+		this.room = room;
+	}
 }
